[89] Refactor pagy exception handling into a concern
