+++ app/controllers/api/public/v1/application_controller.rb
@@ -3,13 +3,9 @@ module API
     module V1
       class ApplicationController < ActionController::API
         include Pagy::Backend
-        include ResponseErrorHandler
+        include ErrorHandlers::Pagy
 
         rescue_from ActiveRecord::RecordNotFound, with: :jsonapi_404
-
-        rescue_from Pagy::OverflowError do |_exception|
-          render_json_error(status: 400, message: I18n.t("pagy.overflow"))
-        end
 
         def jsonapi_404
           render jsonapi: nil, status: :not_found
